Add nessus file import and manual finding adding, asset criticality, tagging
In the finding import form now the engine type can be selected. An engine type 'nessus' is added. The nessus report file then will be parsed and the findings are added via the _import_findings interface.
The nessus has a critical severity in addition to the severities info, low, medium and high. Also most scanning engines uses five severity levels, so a severity level critical is added. In every view, where finding severities are used, this new severity level is added. The color scheme for the severities is updated, so that criticial is now red, high orange, medium yellow, low green and info blue.

A form for adding a finding manually is added.

If there are only three criticality levels, it is most likely, that normally the criticality in the middle is choosen. Therefor, a fourth criticality level ('critical') is added.

The tagging system has been improved. When adding a finding manually, it is possible to add tags in the tag field by pressing the tab-key after entering the tag name. There is also now a edit finding interface, in which it is also possible to add tags.